experiments_v2/requirements.txt
alembic==1.14.0
asttokens==2.4.1
bipartite_learn==0.7.5
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==0.29.33
databricks-sdk==0.37.0
decorator==5.1.1
Deprecated==1.2.15
docker==7.1.0
executing==2.1.0
Flask==3.1.0
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
idna==3.10
imbalanced-learn==0.9.1
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.6
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mlflow==2.17.2
mlflow-skinny==2.17.2
numpy==1.26.4
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.28.3
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.3.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
threadpoolctl==3.5.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.16.0
zipp==3.21.0
